Move marconi workers to library such that they could be tested
